app/requirements.txt
Flask==3.0.0
requests
boto3==1.34.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# Database drivers
PyMySQL==1.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Kafka
kafka-python==2.0.2
avro-python3==1.10.2
